fix: revert unnecessary cast simplification for cross-platform compatibility

+++ src-tauri/src/handlers/bilibili.rs
@@ -1378,7 +1378,9 @@ fn ensure_free_space(target_path: &Path, needed_bytes: u64) -> Result<(), String
                 return Ok(());
             }
             let stat = stat.assume_init();
-            let free_bytes = (stat.f_bavail as u64) * (stat.f_frsize as u64);
+            #[allow(clippy::unnecessary_cast)]
+            #[allow(clippy::useless_conversion)]
+            let free_bytes = u64::from(stat.f_bavail) * stat.f_frsize;
             if free_bytes < needed_bytes {
                 return Err("ERR::DISK_FULL".into());
             }
